requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
dash==2.18.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.16
django-crispy-forms==2.3
django-plotly-dash==2.4.3
dpd-components==0.1.0
Flask==3.0.3
frozenlist==1.4.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
plotly==5.24.1
propcache==0.2.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
retrying==1.3.4
six==1.16.0
sqlparse==0.5.1
tenacity==9.0.0
twilio==9.3.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
